0x8491ab69c89Bb189A8389906c27F016df4A1bF74.json
We have a TetherUS project, with credit/debit cards, and we chose imtoken to handle our 10 million customers. The Club TUS currently has a portfolio of 10,000 customers. and we chose to use this company's wallet for the purpose of moving our customers' resources, our currency is private and we do not want to make it public.

WE NEED TO LIST OUR TOKEN IN THIS WALLET